fixed build (thanks to Markos Chandras)
git-svn-id: file:///home/maarten/qmmp-svn-raw/branches/qmmp-0.2@894 90c681e8-e032-0410-971d-27865f9a5e38

+++ src/plugins/Input/mad/decoder_mad.cpp
@@ -265,7 +265,7 @@ bool DecoderMAD::findHeader()
             {
                 //skip ID3v2 tag
                 uint tagSize = findID3v2((uchar *)stream.this_frame,
-                                         (uint) stream.bufend - (uint) stream.this_frame);
+                                         (ulong) (stream.bufend - stream.this_frame));
                 if (tagSize > 0)
                 {
                     mad_stream_skip(&stream, tagSize);
